session management ready
