+++ webserv_2.1/configuration/default.conf
@@ -35,7 +35,7 @@ server
 	location /cgi-bin
 	{
 		cgi_ext		.py .sh .cgi .rb;
-		cgi_path	cgi /usr/bin/python3 /bin/bash /usr/bin/ruby;
+		cgi_path	cgi /usr/local/bin/python3 /bin/bash /usr/bin/ruby;
 	}
 	location /obj
 	{
