--- webserv_2.1/webpage/account/index.html
@@ -1,47 +1,0 @@
-<!DOCTYPE html>
-<html lang="eng">
-	<head>
-		<meta charset="UTF-8">
-		<meta name="viewport" content="width=device-width, intial-scale=1.0">
-		<title>Login Page</title>
-
-		<!-- Link the external CSS file -->
-		<link rel="stylesheet" type="text/css" href="../css/styles.css">
-	</head>
-
-<body>
-	<header>
-		<!-- Site Name -->
-		<h1>CLJ WEBSERV</h1>
-
-		<!-- Main Menu -->
-		<nav>
-			<ul class="nav-pills">
-				<li><a href="../index.html">Home</a></li>
-				<li><a href="../account/index.html">Account</a></li>
-				<li><a href="../services/index.html">Services</a></li>
-				<li><a href="../gallery/index.html">Gallery</a></li>
-			</ul>
-		</nav>
-	</header>
-
-	<main>
-		<center> <h1> CLJ Login Form </h1> </center>
-		<form action = "../../cgi-bin/acc.py" method = "GET">
-			<div class="container">
-				<label>Username : </label>
-				<input type="text" placeholder="Enter Username" name="username" required>
-				<label>Password : </label>
-				<input type="password" placeholder="Enter Password" name="password" required>
-				<button type="submit">Login</button>
-				No Account?<a href="register.html"> Register Here! </a>
-			</div>
-		</form>
-	</main>
-
-	<footer>
-		<p>&copy; 2023 CLJ Development Team</p>
-	</footer>
-
-</body>
-</html>
