+++ webserv_2.1/webpage/account/register.html
@@ -18,7 +18,7 @@
 		<nav>
 			<ul class="nav-pills">
 				<li><a href="../index.html">Home</a></li>
-				<li><a href="../account/index.html">Account</a></li>
+				<li><a href="../../cgi-bin/account.py">Account</a></li>
 				<li><a href="../services/index.html">Services</a></li>
 				<li><a href="../gallery/index.html">Gallery</a></li>
 			</ul>
@@ -27,7 +27,7 @@
 
 	<main>
 		<center> <h1> CLJ Registration Form </h1> </center>
-		<form action = "../../cgi-bin/acc.py" method = "GET">
+		<form action = "../../cgi-bin/account.py" method = "GET">
 			<div class="container">
 				<label>Username : </label>
 				<input type="text" placeholder="Enter Username" name="username" required>
@@ -36,7 +36,7 @@
 				<label>FirstName : </label>
 				<input type="text" placeholder="Enter FirstName" name="firstname" required>
 				<button type="submit">Register</button>
-				Already Have account?<a href="../../cgi-bin/acc.py"> Login </a>
+				Already Have account?<a href="../../cgi-bin/account.py"> Login </a>
 			</div>
 		</form>
 	</main>
