+++ webserv_2.1/webpage/index.html
@@ -18,7 +18,7 @@
 		<nav>
 			<ul class="nav-pills">
 				<li><a href="index.html">Home</a></li>
-				<li><a href="account/index.html">Account</a></li>
+				<li><a href="../cgi-bin/account.py">Account</a></li>
 				<li><a href="services/index.html">Services</a></li>
 				<li><a href="gallery/index.html">Gallery</a></li>
 			</ul>
